requirements.txt
# Django Framework (4.1.13 is required for djongo compatibility)
Django==4.1.13
djangorestframework==3.14.0

# Database (MongoDB)
djongo==1.3.6
pymongo==3.12.3
# Note: sqlparse version will be resolved automatically by Django

# JWT Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# QR Code Generation
qrcode==7.4.2
Pillow>=10.0.0  # Use latest compatible version with pre-built wheels

# Environment Variables
python-dotenv==1.0.0

# CORS Support
django-cors-headers==4.3.1

# Additional utilities
pytz==2023.3
